requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
torch
torchvision
pillow
requests==2.32.3
python-dotenv==1.0.1
